+++ src/main/java/org/janggo/pseveryday/domain/problem/entity/ProblemTag.java
@@ -1,0 +1,28 @@
+package org.janggo.pseveryday.domain.problem.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "problem_tag")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ProblemTag {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "problem_id")
+    private Problem problem;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "tag_id")
+    private Tag tag;
+
+    public ProblemTag(Problem problem, Tag tag) {
+        this.problem = problem;
+        this.tag = tag;
+    }
+}
