--- src/main/java/org/janggo/pseveryday/problem/repository/TagRepository.java
@@ -1,13 +1,0 @@
-package org.janggo.pseveryday.problem.repository;
-
-import org.janggo.pseveryday.problem.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface TagRepository extends JpaRepository<Tag, Long> {
-    Optional<Tag> findByDisplayName(String displayName);
-    Optional<Tag> findByKey(String key);
-
-    boolean existsByDisplayName(String name);
-}
