--- src/main/java/org/janggo/pseveryday/problem/solvedac/SolvedAcController.java
@@ -1,19 +1,0 @@
-package org.janggo.pseveryday.problem.solvedac;
-
-import lombok.RequiredArgsConstructor;
-import org.janggo.pseveryday.problem.dto.SolvedAcResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/problems")
-public class SolvedAcController {
-    private final SolvedAcService solvedAcService;
-    @GetMapping("/random")
-    public ResponseEntity<SolvedAcResponse.ProblemItem> getRandomProblem() {
-        return ResponseEntity.ok(solvedAcService.getRandomProblem());
-    }
-}
